# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Supernova API
  description: |
    REST API for the Supernova task management application.

    ## Overview
    Supernova is a task app where your lock screen wallpaper IS your task list,
    showing you the ONE thing to do right now.

    ## Authentication
    Most endpoints require JWT authentication via Bearer token in the Authorization header.
    - **Access Token**: 7-day expiry, used for API access
    - **Refresh Token**: 30-day expiry, used to obtain new access tokens
    - **Wallpaper Token**: 365-day expiry, limited scope for widget access

    ## Rate Limiting
    - Global: 100 req/min
    - Auth endpoints: 10 req/min
    - Wallpaper: 20 req/hour
    - Task creation: 30 req/min

  version: 1.0.0
  contact:
    name: Supernova Support
    url: https://github.com/supernova-app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.supernova.app
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/tasks:
    get:
      tags:
        - Tasks
      summary: List all tasks
      description: |
        Get all active tasks for user, sorted by priority score.
        Excludes completed tasks and tasks snoozed until future.
      operationId: listTasks
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: location
          in: query
          description: Context for location-based prioritization
          schema:
            type: string
        - name: time
          in: query
          description: Time context for prioritization
          schema:
            type: string
            enum: [morning, afternoon, evening]
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Tasks
      summary: Create task
      description: |
        Create a new task with natural language parsing.
        Supports patterns like "Buy milk tomorrow at 3pm in 30m"
      operationId: createTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
            examples:
              natural_language:
                summary: Natural language input
                value:
                  rawTitle: Buy groceries tomorrow at 3pm in 30m
              structured:
                summary: Structured input
                value:
                  title: Buy groceries
                  priority: 2
                  context_location: grocery_store
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Parse error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - Tasks
      summary: Delete all tasks
      description: Delete all tasks for user
      operationId: deleteAllTasks
      security:
        - bearerAuth: []
        - {}
      responses:
        '200':
          description: All tasks deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: All tasks deleted
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/tasks/{taskId}:
    get:
      tags:
        - Tasks
      summary: Get task
      description: Get a single task by ID
      operationId: getTask
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - Tasks
      summary: Update task
      description: Update task fields. Sets completed_at/archived_at automatically.
      operationId: updateTask
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/taskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Tasks
      summary: Delete task
      description: Delete a single task
      operationId: deleteTask
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: Task deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/sync:
    post:
      tags:
        - Sync
      summary: Sync changes
      description: |
        Synchronize offline changes with server using last-write-wins conflict resolution.
        Returns all tasks modified since lastSyncAt.
      operationId: syncChanges
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncRequest'
      responses:
        '200':
          description: Sync complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/sync/status:
    get:
      tags:
        - Sync
      summary: Get sync status
      description: Get sync status and metadata
      operationId: getSyncStatus
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Sync status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================
  # WALLPAPER ENDPOINT
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  parameters:
    taskId:
      name: taskId
      in: path
      required: true
      description: Task UUID
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              message: Authentication required
              code: AUTHENTICATION_ERROR
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
        due_date:
          type: string
          format: date
          nullable: true
        due_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
          nullable: true
        estimate_minutes:
          type: integer
          nullable: true
        priority:
          type: integer
          minimum: 1
          maximum: 3
          description: "1=critical, 2=normal, 3=low"
        completed:
          type: boolean
        completed_at:
          type: string
          format: date-time
          nullable: true
        snoozed_until:
          type: string
          format: date-time
          nullable: true
        context_location:
          type: string
          nullable: true
        context_time:
          type: string
          enum: [morning, afternoon, evening]
          nullable: true
        energy_required:
          type: string
          enum: [low, medium, high]
          nullable: true
        times_rescheduled:
          type: integer
        original_due_date:
          type: string
          format: date
          nullable: true
        x:
          type: number
          nullable: true
          description: Spatial coordinate
        y:
          type: number
          nullable: true
          description: Spatial coordinate
        is_subtask:
          type: boolean
        is_recurring:
          type: boolean
        echo_interval:
          type: string
          nullable: true
          description: "Recurrence interval (1d, 1w, etc.)"
        archived:
          type: boolean
        archived_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        score:
          type: integer
          description: Calculated priority score (higher = more urgent)
    CreateTaskRequest:
      type: object
      properties:
        rawTitle:
          type: string
          description: Natural language task input
        title:
          type: string
          description: Plain title (alternative to rawTitle)
        priority:
          type: integer
          minimum: 1
          maximum: 3
        context_location:
          type: string
        context_time:
          type: string
          enum: [morning, afternoon, evening]
        x:
          type: number
        y:
          type: number
        is_subtask:
          type: boolean
        is_recurring:
          type: boolean
        echo_interval:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        priority:
          type: integer
          minimum: 1
          maximum: 3
        estimate_minutes:
          type: integer
        due_date:
          type: string
          format: date
        due_time:
          type: string
        completed:
          type: boolean
        snoozed_until:
          type: string
          format: date-time
        context_location:
          type: string
        context_time:
          type: string
          enum: [morning, afternoon, evening]
        energy_required:
          type: string
          enum: [low, medium, high]
        x:
          type: number
        y:
          type: number
        is_subtask:
          type: boolean
        is_recurring:
          type: boolean
        echo_interval:
          type: string
        archived:
          type: boolean
        source:
          type: string
          description: "Set to 'widget' for widget-based completions"

    # ----------------------------------------
    # Sync Schemas
    # ----------------------------------------
    SyncRequest:
      type: object
      properties:
        lastSyncAt:
          type: integer
          description: Unix timestamp (ms) of last sync
        pendingChanges:
          type: array
          items:
            $ref: '#/components/schemas/SyncChange'
    SyncChange:
      type: object
      required:
        - type
        - clientId
        - timestamp
      properties:
        type:
          type: string
          enum: [create, update, delete]
        action:
          type: string
          enum: [create, update, delete]
          description: Legacy alias for type
        clientId:
          type: string
          description: Client-generated UUID
        timestamp:
          type: integer
          description: Unix timestamp (ms)
        data:
          type: object
          description: Change payload (task data)
    SyncResponse:
      type: object
      properties:
        syncedAt:
          type: integer
          description: Server timestamp (ms)
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        results:
          type: object
          properties:
            applied:
              type: integer
            rejected:
              type: integer
            conflicts:
              type: array
              items:
                $ref: '#/components/schemas/SyncConflict'
        conflicts:
          type: array
          items:
            $ref: '#/components/schemas/SyncConflict'
          description: For backward compatibility
    SyncConflict:
      type: object
      properties:
        clientId:
          type: string
        reason:
          type: string
          enum: [already_exists, not_found, stale_data, server_error]
        serverData:
          $ref: '#/components/schemas/Task'
    SyncStatus:
      type: object
      properties:
        taskCount:
          type: integer
          description: Total active tasks
        lastModified:
          type: integer
          nullable: true
          description: Timestamp of most recent task update
        serverTime:
          type: integer
          description: Current server timestamp

    # ----------------------------------------
    # Error Schemas
    # ----------------------------------------
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - AUTHENTICATION_ERROR
                - AUTHORIZATION_ERROR
                - NOT_FOUND
                - CONFLICT
                - RATE_LIMIT
                - DATABASE_ERROR
                - CACHE_ERROR
                - INTERNAL_ERROR
            timestamp:
              type: string
              format: date-time
            path:
              type: string
            method:
              type: string
            details:
              type: object
    ValidationError:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
              msg:
                type: string
              path:
                type: string
              location:
                type: string